Add operator onChangeFlagOnFail.
It seems like `onChange` except that if property y fails, a flag file
is generated. On next runs, if the flag file is present, property y is
executed even if property x doesn't change.

With `onChange`, if y fails, the property x `onChange` y returns
`FailedChange`. But if this property is applied again, it returns
`NoChange`. This behavior can cause trouble...