add Maintainer entries for contributed modules
These show up in the haddock documentation on the sidebar.
The author emails are not hyperlinked, so hopefully this is not a spam
source.

Keeping track of Maintainers of modules is becoming necessary because I
don't use all these modules and am not the best person to maintain them,
beyond simple changes to keep them building. I'll loop in the Maintainer if
there's a bug etc on their module.