change HostContext for containers
Privdata that uses HostContext inside a container will now have the name of
the container as its context, rather than the name of the host(s) where the
container is used. This allows eg, having different passwords for a user in
different containers.

Note that previously, propellor would prompt using
the container name as the context, but not actually use privdata using that
context; so this is a bug fix.

I don't entirely like the implementation; I had to put the code to change
the context in PropAccum, and it's not generalized past PrivInfo.