Add simplistic support for apt-software-properties-common.
I know it's definitely a very Buntish thing, but for adding some larger
collections of software on a very Buntish box, it definitely comes in
handy.

I also think that it might be possible to support a lot of the features
using propellor properties directly as a fallback, and that might be
something I could look into as well.

Thanks!