hasGroupCache -> hasCache
I was originally thinking that the name `Ccache.hasCache` might be for a
property `User -> Property DebianLike`.  However, someone wanted to
write a property configuring a user cache, it would probably have the
standard location `~/.ccache`.  This cache would be implicitly created
when required, so the name `Ccache.hasCache` would be needed.