throw StopPropellorException if unmounting fails
It would be bad if unmount somehow failed and then another property tried
to reformat the mounted device or rm -rf the mount point. While I don't see
anything that does that, and while lazy umount can't fail as far as I know,
it was throwing an exception that stopped propellor and let's cautiously
keep it that way.