Makefile: Removed "run" target which was default target.
"make" now only builds propellor, does not run it.

Note that propellor 1.0.0 and earlier relied on this target for the
Cron.runPropellor property's cronjob to work, so upgrading directly to this
version from 1.0.0 would break that cron job.

(Debian has 0.9.1 in stable, and 3.2.3 will be in the next stable
release. Upgrading between Debian releases is supposed to be done
without skipping a release, so that's ok.)