Implemented hostChroot
As originally seen in my slides at Linux.Conf.Au 2017 in January.

Now that it's not vaporware, it allows one Host to build a disk image that
has all the properties of another Host.

It was easier than I thought to implement this! As expected, Info
propagation was slightly tricky. Also, I originally had a lot of machinery
to try to use Info to detect infinitely nested chroot loops. But, my
machinery didn't work, and when I tested it, ghc did a much better job,
causing a "warning: <<loop>>" message to be output instead of such a
property using infinite disk space.

This commit was sponsored by Bruno BEAUFILS on Patreon.