Added Monoid instances for Property and RevertableProperty.
* Added Monoid instances for Property and RevertableProperty.
* Removed applyToList. Instead, use mconcat. (API change)
  Eg, if you had: applyToList accountFor [User "joey", User "root"]
  use instead: mconcat (map accountFor [User "joey", User "root"])

mappend x y is basically the same as x `before` y. In particular, if x
fails to be ensured, it won't ensure y. This seems to make sense, since
applyToList had that behavior, and so does the Monoid for Propellor
Result. The alternative would be to try to ensure both and combine
the results. However, I don't see any reason to do it that way.

It would be nice if the description of both properties were displayed when
ensuring the combination. But then, it would need to display eg:

	ensuring x..ok
	ensuring y..failed
	ensuring x and ensuring y..failed

Without a way to get rid of that redundant last line, I don't want to do
that.

Note that the haddocks for the Monoid instances need a really wide screen
to display! This is IMHO an infelicity in haddock, and I can't do anything
about it really.

This commit was sponsored by Fernando Jimenez on Patreon.