fix diskimage finalization bug
  * Fix bug when using setContainerProps with a chroot that prevented
    properties added to a chroot that way from being seen when propellor
    was running inside the chroot. This affected disk image creation, and
    possibly other things that use chroots.

The problem was, propagateChrootInfo was being passed the initial
version of the Chroot, but then the Chroot got more properties
added, and so those were not recorded in the _chroot info.

Fix was simply to make InfoPropagator be passed the Chroot as an
additional parameter, so Chroot.provisioned' can pass in the final
Chroot to it.