avoid buffering container chain output
When provisioning a container, output was buffered until the whole process
was done; now output will be displayed immediately.

I know this didn't used to be a problem. I belive it was introduced by
accident when propellor started using concurrent-output. I know I've seen
it for a while and never was bothered enough to get to the bottom of it;
apparently "a while" was longer than I thought.

Also refactored code to do with chain provisioning to all be in
Propellor.Engine and avoided some duplication.

This commit was sponsored by Anthony DeRobertis on Patreon.