DiskImage: Removed grubBooted
Properties that used to need it as a parameter now look at Info about the
bootloader that is installed in the chroot that the disk image is created
from. (API change)

This is a simplication, and avoids the user needing to repeat themselves
in the propellor config, thus avoiding mistakes.

When no boot loader is installed, or multiple different ones are,
disk image creation will fail, which seems reasonable.

This commit was sponsored by Jake Vosloo on Patreon.