disable buffering earlier
The "fatal: Couldn't find remote ref HEAD" persists, and is intermittent
so hard to be sure but it seemed that disabling buffering earlier
avoided it. Now done first thing on start.

I was not able to find anything that reads from stdin other than
getMarked, but perhaps there is something..