use stack build --dry-run to check if deps are installed
The system might have stack installed, so stack --version would succeed,
but need stack setup to be run to get a working ghc.

This will fail if stack is not installed, or if stack setup needs to be
run. It does not fail when haskell libraries are not installed, but
stack build will install those anyway, so that's ok.