add Propellor.Property.Versioned
New module which allows different versions of a property or host to be
written down in a propellor config file. Has many applications,
including staged upgrades and rollbacks.

Note that it currently only supports RevertableProperty that has the same
metatypes for its setup and cleanup sides. And, each RevertableProperty in
a version definition needs to have the same metatypes as the others too.
I tried a couple of times to add support for differing metatypes, but
it got beyond my avilities to do.

This commit was sponsored by Jeff Goeke-Smith on Patreon.