Grub.bootsMounted: Avoid failing when proc sys etc are already mounted within the chroot.
Perhaps it should keep track of what was mounted before and restore it,
but it would complicate it a lot, and I doubt it would be needed by any use
of this property. Usually, this property will come after a chroot
provisioning property, which may leave proc and sys mounted, but that's
done only so that the chroot can later be used; and when bootsMounted is
applied to a chroot, the goal is presumably to boot the underlying disk
soon.