Make addInfo accumulate Info in order properties appear, not reverse order
This fixes a bug involving reverting Systemd.resolvConfed or
Systemd.linkJournal.

addInfo was prepending to the list for efficiency. But, that was in
conflict with mappend of two Info, which appended the second to the first.
In the case where Systemd.resolvConfed was added reverted, to override the
one added by default, that led to a list of info that had first the
reversion and then the default, so the default won. Which was wrong.

So, make addInfo accumulate in the same order mappend combines things, even
though it's a little less efficient. The efficiency probably does not
matter; there is not typically a whole lot of info.

There's some risk this change has unexpected consequences, if something
relied on the old addInfo order without using fromInfo to access to info.
But if so, that something would have been broken before when two Info
properties were combined. With this change, it would just be broken the
other way around.

This commit was sponsored by Jochen Bartl on Patreon.