don't start services in chroots
Service: Changed to use invoke-rc.d rather than service for starting
services. This notably means that in chroots, services will not be started.

This seems like the right thing to do, because running services inside
a chroot is problimatic. Use a real container system for running services.

Fixes the problem that, when building a disk image, using eg,
Apt.servicesInstalledRunning caused the service to be started in the
chroot, even though a policy-rc.d was in place to prevent it, because the
"service" command does not check policy-rc.d.

This commit was sponsored by John Peloquin on Patreon.