Apt.installedBackport replaced with Apt.backportInstalled
Apt.installedBackport would do this:

    apt-get install -t stretch-backports foo bar

Apt.backportInstalled does this:

    apt-get install foo/stretch-backports bar/stretch-backports

The Apt.installedBackport behaviour can install the dependencies of foo and bar
from stretch-backports even when the versions in stretch will satisfy the
dependencies of the backports of foo and bar.  So this property can result in
very many more backports being installed on the host when intended.  But the
number of installed backports should always be minimised.

Worse, whether this happens is highly dependent on the system state, and the
order in which other properties get ensured.  For example,

    & Apt.installed ["dgit"]
    & Apt.installedBackport ["dgit"]

will install only dgit from stretch-backports, but unless debhelper and
devscripts happen to already be installed,

    & Apt.installedBackport ["dgit"]
    & Apt.installed ["dgit"]

will install dgit, debhelper, devscripts and maybe more from backports.  This is
surprising, difficult to debug, and breaks the expectation that when the order
in which properties are ensured is not specified with connectives like
`requires` and `before`, ensuring them in any order will produce the same
result.

Property renamed because user configs should not silently break, as they would
if they did not list dependencies that must be installed from stable-backports.

Signed-off-by: Sean Whitton <spwhitton@spwhitton.name>