modernized and simplified the MetaTypes implementation
now that compatability with ghc 7 is no longer needed.

Data.Type.Bool contains effectively the same stuff that was implemented
here, so removed my code.

Tried to use Data.Type.Equality instead of my EqT, but it seems to be some
other type of (type level) equality, and didn't compile. Instead went with
the simpler EqT implementation that newer ghc versions allow.

The rest of the changes are simply better syntax for defining type
families.

And upon using that syntax, ghc noticed that `type family a + b`
does not have kind "ab" like I wrote before, but is kind *.

Tested on debian stable with ghc 8.0.1.

This commit was sponsored by John Pellman on Patreon.