Split mailname property out of Hostname.sane
Since bad mailname guesses can lead to ugly surprises. (API change)

Kept it in the Hostname module for easy discoverability, and similar to
Hostname.searchDomain it sets a value based on the hostname so makes sense
to keep it in that module.

Didn't implement the mailname equivilant of Hostname.setTo, because it's
trivial to write the mailname file with a custom value if desired.

This commit was sponsored by John Pellman on Patreon.