remove boot part
uboot only needs to be in the boot sector of the sd card and can read
the kernel from sata, so a separate boot partition on the sd card is not
needed.

I am not going to automate creating the sd card with uboot in
the boot sector with propellor, but it's a simple dd to set up.

This way, crappy sd card fsck problems will not prevent booting..