libghc-stm-dev package won't be in new versions of debian
In Bootstrap, still try to install the package to support bootstrapping
to old systems. If it's not available the bootstrapping will still
succeed. The added apt-cache check is only to avoid apt complaining when
asked to install a not available package.

In debian/control, depend on ghc that includes stm, although propellor
still supports being used with older versions of ghc
Of course this control file is not being used for the package in debian
any longer afaik, so something else will be done there.