fix withOS type level bug
withOS had a type level bug that allowed ensureProperty to be used inside
it with a Property that does not match the type of the withOS itself.

Propellor.Property.Cron.runPropellor is a Property DebianLike; it was
incorrectly a Property UnixLike before and that wrong type was hidden by
the withOS bug.

This commit was sponsored by Jack Hill on Patreon.