sbuild: break Apt.stdSourcesList out of Sbuilt.built
Apt.stdSourcesList supports only Debian, so having it inside Sbuilt.built was
blocking using Sbuild.built with non-Debian schroots.

Note that we cannot break out the installation of eatmydata and ccache, because
Sbuild.built assumes that those are available in the schroot config files that
it writes out.

This is a minimal fix.  If it becomes clear that there are properties which
would be wanted in every Buntish schroot, say, we might replace
Sbuild.osDebianStandard with an Sbuild.standard property which uses pickOS to
install the right set of properties.

Reported-by: David Bremner <david@tethera.net>
Signed-off-by: Sean Whitton <spwhitton@spwhitton.name>