Fix bug in File.containsShellSetting
It replaced whole shell conffile content with the setting if the file did not
previously contain a line setting the key to some value.

InsertSection is passed the whole content of the file, which does not
contain the section yet (here the section is a single line).

Probably this was not noticed because it's common for a shell config file
to contain default values, so the buggy InsertSection rarely gets used.

This commit was sponsored by Ethan Aubin.